Profiling tests for P2
The photos are on my web folder in sigma
WARNING
These tests are long and take long so, in order to save time,
their iterators are set to 1 as the upper limit. If you want to run the tests
see the comment right next to the variable to set the real values.